rename routetable resource to cproutetables
