+++ service/controller/clusterapi/v29/cluster_resource_set.go
@@ -21,6 +21,7 @@ import (
 	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/bridgezone"
 	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/cpf"
 	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/cpi"
+	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/cproutetables"
 	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/cpvpccidr"
 	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/ebsvolume"
 	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/encryption"
@@ -32,7 +33,6 @@ import (
 	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/natgatewayaddresses"
 	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/peerrolearn"
 	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/region"
-	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/routetable"
 	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/s3bucket"
 	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/s3object"
 	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/resource/secretfinalizer"
@@ -464,15 +464,15 @@ func NewClusterResourceSet(config ClusterResourceSetConfig) (*controller.Resourc
 		}
 	}
 
-	var routeTableResource controller.Resource
+	var cpRouteTablesResource controller.Resource
 	{
-		c := routetable.Config{
+		c := cproutetables.Config{
 			Logger: config.Logger,
 
 			Names: strings.Split(config.RouteTables, ","),
 		}
 
-		routeTableResource, err = routetable.New(c)
+		cpRouteTablesResource, err = cproutetables.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -580,7 +580,7 @@ func NewClusterResourceSet(config ClusterResourceSetConfig) (*controller.Resourc
 		accountIDResource,
 		natGatewayAddressesResource,
 		peerRoleARNResource,
-		routeTableResource,
+		cpRouteTablesResource,
 		vpcCIDRResource,
 		tccpOutputsResource,
 		tccpSubnetResource,
