+++ service/controller/clusterapi/v29/resource/cproutetables/create.go
@@ -1,4 +1,4 @@
-package routetable
+package cproutetables
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 )
 
 func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
-	err := r.addRouteTableMappingsToContext(ctx)
+	err := r.addRouteTablesToContext(ctx)
 	if err != nil {
 		return microerror.Mask(err)
 	}
