+++ service/controller/clusterapi/v29/resource/cproutetables/delete.go
@@ -1,4 +1,4 @@
-package routetable
+package cproutetables
 
 import (
 	"context"
