+++ service/controller/clusterapi/v29/resource/cproutetables/error.go
@@ -1,4 +1,4 @@
-package routetable
+package cproutetables
 
 import (
 	"github.com/giantswarm/microerror"
